Small bug fix

+++ nnlibrary/utils/loss.py
@@ -36,8 +36,8 @@ class CrossEntropyLoss(nn.CrossEntropyLoss):
         label_smoothing: float = 0
     ) -> None:
         device = 'cuda' if torch.cuda.is_available() else 'cpu'
-        if isinstance(weight, list): weight = torch.as_tensor(weight, device=device).long()
-        elif isinstance(weight, Tensor): weight = weight.to(device=device).long()
+        if isinstance(weight, list): weight = torch.as_tensor(weight, device=device)
+        elif isinstance(weight, Tensor): weight = weight.to(device=device)
         else: raise ValueError(f"Weight type must be a List or Tensor, current type: {type(weight)}")
         super().__init__(weight, size_average, ignore_index, reduce, reduction, label_smoothing)
